feat: add "candidate" branch to protection rule for all envs
Due to the way we're using Github Actions to respond to issues,
any events in response to issue comments are run in the context
of the candidate branch, which means that they were previously
unable to access secrets from other environments attached to
other branches.

This change ensures that all environments created can be
accessed by their matching branch, but also the candidate branch.