chore: rename 'snaps' to 'repos' in config
Looking at how we're actually using this tool, a lot of the
configuration is actually about 'repos', not 'snaps'. In the
snapcrafters org, most of the repos are named after the snaps
themselves, but not in cases where one repo might represent
multiple snaps.

This change renames the top level config field, and some of
the internal variables ready for enabling multi-snap tokens.